Split myworker.worker.ts and pull in supernote-typescript's tree-shaking fix
Bumps the supernote-typescript submodule to pick up
https://github.com/philips/supernote-typescript/pull/46 (fixes issue #45):
adding "sideEffects": false to its package.json lets esbuild actually
tree-shake through its barrel index.ts, instead of conservatively assuming
any module the barrel touches could have side effects and keeping all of
it. That alone shrank the standalone web component bundle from 3.84MB to
2.38MB.

Rebuilding after that bump showed pdf-lib was still fully present in the
bundle - a second, separate source of bloat: myworker.worker.ts
unconditionally imported createPdfContext/addPdfPage (PDF-building
functions) at module scope, even though the web component's ImageConverter
only ever sends 'convert' (rasterization) messages, never 'buildPdf'. Since
esbuild-plugin-inline-worker bundles a worker file as one self-contained
blob regardless of which message types a given consumer actually uses,
pdf-lib came along into the standalone bundle no matter what.

Split the one worker into two:
- rasterize.worker.ts: rasterization only ('convert'), shared by
  SupernoteView, SupernoteEmbed, and <supernote-viewer> (all through
  src/render/imageConverter.ts) - stays free of any PDF-related import.
- pdfBuild.worker.ts: PDF assembly only ('buildPdf'), used solely by
  main.ts's buildPdfInWorker() for the plugin's own PDF export feature.

Net effect on the standalone bundle: 3.84MB -> 923KB (down ~76%), with
pdf-lib now fully absent (0 references, confirmed by grep) - verified
end-to-end in a real headless Chromium that rasterization still works
correctly with the split worker. main.js (the Obsidian plugin bundle,
which needs both workers) grew slightly, an acceptable trade since it was
never the size-sensitive artifact.

Also: worker.d.ts now declares both new bare *.worker specifiers (tried a
single wildcard `declare module '*.worker'` pattern first - TypeScript's
wildcard module matching doesn't apply to bare, slash-free specifiers like
these, so kept two explicit declarations instead). Updated stale
myworker.worker.ts references in AGENTS.md and CLAUDE.md.
